SideBarLayout moved to where it is used
